Add new assets for game including images, animations, and metadata
- Added new PNG assets: hud_tileset.png, moon.png, pillar.png, sandblock.png, simon_walk.png, and torch.png.
- Created corresponding text files for each asset to define their states and dimensions.
- Introduced sequence files for simon_walk and torch animations to manage their respective frames.
- Updated sandblock asset to include multiple states for varied gameplay.

+++ samples/Benchmark.c
@@ -29,7 +29,7 @@ int main(void) {
 
   /* setup engine */
   TLN_Init(HRES, VRES, 1, NUM_SPRITES, 0);
-  framebuffer = malloc(HRES * VRES * 4);
+  framebuffer = malloc((size_t)HRES * VRES * 4);
   TLN_SetRenderTarget(framebuffer, HRES * 4);
   TLN_DisableBGColor();
 
@@ -101,7 +101,7 @@ static uint32_t Profile(void) {
 
   t0 = TLN_GetTicks();
   do {
-    TLN_UpdateFrame(frame);
+    TLN_UpdateFrame((int)frame);
     frame++;
   } while (frame < NUM_FRAMES);
   elapse = TLN_GetTicks() - t0;
